Use xmlrpc in get_name_from_nvr() in rpminspect_runner.sh
Rather than using sed to break up the nvr string, ask the Koji hub for
information about that build.  It knows the package name, so extract
it from the XML-RPC result.

This is not as ideal as I would like, but it is talking to Koji.  What
would be nice is to submit a PR to Koji that extends the buildinfo
command to provide Name, Version, Release, Epoch, Arch all as separate
fields.  I will do that later.

Signed-off-by: David Cantrell <dcantrell@redhat.com>