Run all inspections at once and cache results
This fundamentally changes how we run rpminspect and how we treat its
results, so the patch is huge...

Previous approach when we were running rpminspect separately for each
inspection was problematic as rpminspect was unpacking sources on every
invocation. Which was perfectly fine in 99% of cases, but some builds
can be 100+ gigs unpackaged... And unpacking is CPU-bound operation, and
machines in CI sometimes don't have enough... so CI would kill
rpminspect because it was taking too long.. etc. You get the point.

So here, we are going to run rpminspect only once, and then return cached
results when asked for specific inspection.

This is also getting super-painful to maintain so it's probably time to
rewrite the whole script in Python.