Clean up around RPMINSPECT_WORKDIR
This was confusing as it wasn't actually doing
what one was expecting it to be doing.
rpminspect was not using it as its workdir, because
it was not called with --workdir param.
Instead, the directory was only used as a cache for results.

Also, tmt ignores the container workdir altogether and instead
calls the runner script in its own directory.

This commit makes the workdir relative to $PWD and tells
rpminspect to actually use it.

Signed-off-by: Michal Srb <michal@redhat.com>