Make ClamAV virus database updates more robust
See: OSCI-6251

freshclam will try to download incremental patches, but if one
or more patches are missing, it will update the database only
partially. If we then try to run freshclam again, it will try
to download the whole daily.cvd file.

If even the second run fails, we will continue with the older
version of the database.

Signed-off-by: Michal Srb <michal@redhat.com>