Don't skip the before build lookup if the after build NVR was provided
Previously, if the runner would be called with the KOJI_NVR env var set,
the before build lookup section would be skipped.

In the past, we were only setting KOJI_NVR for modules and we don't
expect any before build lookups there. However, with Konflux builds
(imported to Brew as draft builds), we started setting KOJI_NVR for RPM
builds as well.

This commit should fix the problem. And hopefully make the code slightly
more readable as well.

Signed-off-by: Michal Srb <michal@redhat.com>